int64_t instead of size_t

+++ kizhin.evgeniy/S1/type_utils.hpp
@@ -1,12 +1,13 @@
 #ifndef SPBSPU_LABS_2025_AADS_A_KIZHIN_EVGENIY_S1_TYPE_UTILS_HPP
 #define SPBSPU_LABS_2025_AADS_A_KIZHIN_EVGENIY_S1_TYPE_UTILS_HPP
 
+#include <cstdint>
 #include <iosfwd>
 #include <utility>
 #include "forward-list-fwd-declaration.hpp"
 
 namespace kizhin {
-  using Numbers = ForwardList< std::size_t >;
+  using Numbers = ForwardList< std::int64_t >;
   using SequenceT = std::pair< std::string, Numbers >;
 }
